Move BarrelCompartmentEntity.java to a subpackage for vanilla compartments to help prepare for the eventual break off into a separate mod we depend on for the TFC specific things
