+++ src/main/java/com/alekiponi/firmaciv/common/entity/vehiclehelper/compartment/vanilla/package-info.java
@@ -1,0 +1,9 @@
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+@FieldsAreNonnullByDefault
+package com.alekiponi.firmaciv.common.entity.vehiclehelper.compartment.vanilla;
+
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
